cambio la distancia a double

+++ src/test/java/ar/edu/utn/frba/dds/impactoambiental/models/BaseTest.java
@@ -121,7 +121,7 @@ public abstract class BaseTest extends AbstractPersistenceTest
     return new Parada("Parada", crearDistanciaEnKm(distanciaAAnterior), crearDistanciaEnKm(distanciaAProxima));
   }
 
-  protected Distancia crearDistanciaEnKm(int kilometros) {
+  protected Distancia crearDistanciaEnKm(double kilometros) {
     return new Distancia(kilometros, Unidad.KM);
   }
